backend/pyproject.toml
[tool.poetry]
name = "hr-backend"
version = "0.1.0"
description = "TalentLens HR backend service (FastAPI)"
authors = ["HR Team <hr@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = {version = "^0.30.0", extras = ["standard"]}
pydantic = "^2.7.0"
sqlmodel = "^0.0.16"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
python-multipart = "^0.0.20"
pdfplumber = "^0.11.8"
python-docx = "^1.2.0"
aiofiles = "^25.1.0"
openpyxl = "^3.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"
ruff = "^0.6.9"


[[tool.poetry.source]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
